Added MainMenu and GameOver scenes
MainMenu scene is first to load and it can start the game to level1 and quit currently.
Game over scene can reload to the main menu or quit.
Game does load GameOver scene when hp is below 0.
Bug currently where mouse is not able to be used on Game Over screen after dying.